Function oak::channel_create_with_label

pub fn channel_create_with_label(
    label: &Label
) -> Result<(WriteHandle, ReadHandle), OakStatus>

Create a new unidirectional channel.

The provided label must be equal or more restrictive than the label of the calling node, i.e. the label of the calling node must "flow to" the provided label.

On success, returns WriteHandle and a ReadHandle values for the write and read halves (respectively).