Function oak::run_event_loop

pub fn run_event_loop<T: Decodable, N: Node<T>>(node: N, in_channel: ReadHandle)

Run an event loop on the provided node:

wait for new messages on the provided channel in_channel

if the runtime signals that the Node was terminated while waiting, then exit the event loop

otherwise, read the available message via the provided channel handle

decode the message from (bytes + handles) to the specified type T

pass the typed object to the Node::handle_command method of the node, which executes a single iteration of the event loop

Note the loop is only interrupted if the Node is terminated while waiting. Other errors are just logged, and the event loop continues with the next iteration.