Function oak::grpc::event_loop

pub fn event_loop<T: OakNode>(
    node: T, 
    grpc_in_handle: ReadHandle
) -> Result<(), OakStatus>

Perform a gRPC event loop for a Node.

Invoking the given node's invoke method for each incoming request that arrives on the inbound channel as a serialized GrpcRequest message, giving the invoke method the outbound channel for encapsulated responses to be written to.