Trait oak::grpc::ServerNode

pub trait ServerNode {
    fn invoke(
        &mut self, 
        method: &str, 
        req: &[u8], 
        writer: ChannelResponseWriter
    );
}

Trait for Oak Nodes that act as a gRPC services.

A ServerNode instance is normally passed to oak::run_command_loop, to allow repeated invocation of its invoke() method.

You can choose to implement this trait yourself, or use the convenient Dispatcher created by the gRPC code generator.

Required methods

fn invoke(&mut self, method: &str, req: &[u8], writer: ChannelResponseWriter)

Process a single gRPC method invocation.

The method name is provided by method and the incoming serialized gRPC request is held in req. Response messages should be written using writer.write, followed by writer.close.