Module oak::handle

Utilities for visiting, extract and injecting handles.

Applications will usually not interact with the types in this module directly, as the HandleVisit trait is automatically derived for all protobuf types compiled with oak_utils, and extracting and injecting handles is taken care of by oak::io::Receiver and oak::io::Sender.

Structs

Receiver

Receiver handle for an Oak channel. This type is sent over the wire, but in types generated by prost it is replaced with oak::io::Receiver.

Sender

Sender handle for an Oak channel. This type is sent over the wire, but in types generated by prost it is replaced with oak::io::Sender.

Traits

HandleVisit

Visit all handles present in a type.

Functions

extract_handles

Return all handles in T.

inject_handles

Inject handles into a message.

invalid

Returns an intentionally invalid handle.

is_valid

Check this handle is valid.

Derive Macros

HandleVisit

Automatically derives the HandleVisit trait for structs and enums generated by prost.