Trait oak::handle::HandleVisit

pub trait HandleVisit {
    fn fold<B>(&mut self, init: B, f: fn(B, &mut u64) -> B) -> B;
}

Visit all handles present in a type.

This trait can be automatically derived:

use oak_io::{handle::HandleVisit, Handle};

#[derive(HandleVisit)]
struct Thing {
    handle: Handle,
}

Alternatively, you can implement the trait manually. This is required if you want the trait to visit handles that are directly contained in a type.

use oak_io::{handle::HandleVisit, Handle};

struct Thing {
    handle: Handle,
}

impl HandleVisit for Thing {
    fn fold<B>(&mut self, init: B, mut f: fn(B, &mut Handle) -> B) -> B {
        f(init, &mut self.handle)
    }
}

Required methods

fn fold<B>(&mut self, init: B, f: fn(B, &mut u64) -> B) -> B

Invokes the provided closure on every handle contained in self.

The mutable reference allows modifying the handles.