Function oak::io::entrypoint_node_create

pub fn entrypoint_node_create<T: WasmEntrypoint<Message = InitWrapper<Init, Command>>, Init: Encodable + Decodable, Command: Encodable + Decodable>(
    name: &str, 
    label: &Label, 
    wasm_module_name: &str, 
    init: Init
) -> Result<Sender<Command>, OakError>

Creates a node and corresponding inbound channel of the same type, sends an init message to it, and returns a Sender of command messages; for nodes that are instantiated via the crate::entrypoint_command_handler_init macro.