Function oak::io::forward_invocation

pub fn forward_invocation(
    invocation: Invocation, 
    invocation_sender: &Sender<Invocation>
) -> Result<()>

Sends crate::grpc::Invocation (containing oak_io::Receiver and oak_io::Sender) through invocation sender if oak_io::Receiver label flows to invocation sender's label. If failed - sends error back through oak_io::Sender.

Useful for sending invocations from router nodes and checking label correctness without actually reading the contents of invocation.