Struct oak::io::Channel

pub struct Channel {
    pub handle: WriteHandle,
}

Wrapper for a WriteHandle to implement the std::io::Write trait.

Methods taking Write trait objects require mutable references, so use a distinct type rather than implementing the trait on the base WriteHandle type.

Fields
handle: WriteHandle
Methods

impl Channel

pub fn new(handle: WriteHandle) -> Self

Create a new io::Channel that uses the given channel handle.

pub fn close(self) -> Result<()>

Close the underlying channel handle.