Struct oak::io::Receiver

pub struct Receiver<T: Decodable> {
    pub handle: ReadHandle,
    // some fields omitted
}

Wrapper for a handle to the read half of a channel, allowing to receive data that can be decoded as bytes + handles via the Decodable trait.

For use when the underlying Handle is known to be for a receive half.

Fields
handle: ReadHandle
Implementations

impl<T: Decodable> Receiver<T>

pub fn new(handle: ReadHandle) -> Self

pub fn close(&self) -> Result<(), OakStatus>

Close the underlying channel used by this receiver.

pub fn receive(&self) -> Result<T, OakError>

Attempt to wait for a value on this receiver, blocking if necessary.

See https://doc.rust-lang.org/std/sync/mpsc/struct.Receiver.html#method.recv

pub fn try_receive(&self) -> Result<T, OakError>

Attempt to read a value from this receiver, without blocking.

See https://doc.rust-lang.org/std/sync/mpsc/struct.Receiver.html#method.try_recv

pub fn wait(&self) -> Result<ChannelReadStatus, OakStatus>

Wait for a value to be available from this receiver.

Returns ChannelReadStatus of the wrapped handle, or Err(OakStatus::ErrTerminated) if the Oak Runtime is terminating.

impl<T: Decodable> Receiver<T>

pub fn as_proto_handle(&self) -> Receiver