Macro oak::entrypoint

macro_rules! entrypoint {
    ($name : ident < $msg : ty > => $handler : expr) => { ... };
}

Register a new Node entrypoint.

This registers the entrypoint name and the expression that runs an event loop, and it includes the static type of messages that are read from the inbound channel.

The first node of an Oak application usually receives a ConfigMap message from the Oak Runtime via the initial incoming channel, which is reflected in the type of message associated with the entrypoint.

use oak::io::{Receiver, ReceiverExt};
use oak_abi::proto::oak::application::ConfigMap;

#[derive(Default)]
struct DummyNode;

impl oak::CommandHandler for DummyNode {
    // ...
}

oak::entrypoint!(dummy<ConfigMap> => |receiver: Receiver<ConfigMap>| {
    let config_map = receiver.receive().unwrap();
    let grpc_server_listen_address = String::from_utf8(
        config_map
            .items
            .get("grpc_server_listen_address")
            .unwrap_or(&"[::]:8080".as_bytes().to_vec())
        .to_vec()).unwrap();
    let grpc_channel = oak::grpc::server::init(&grpc_server_listen_address)
        .expect("could not create gRPC server pseudo-node");
    let dispatcher = DummyNode::default();
    oak::run_command_loop(dispatcher, grpc_channel.iter());
});

Entrypoints need to be declared in the global scope (as opposed to for example in a function body). Is recommended but not required to define all entrypoints of your module in lib.rs.

If instantiating your Node requires some setup, it is possible to do that in the Node expression too:

use oak::io::ReceiverExt;
use oak_abi::proto::oak::application::ConfigMap;

oak::entrypoint!(its_complicated<ConfigMap> => |_receiver| {
    init_all_the_things();
    let dispatcher = DummyNode::default();
    let grpc_channel = oak::grpc::server::init("[::]:8080")
        .expect("could not create gRPC server pseudo-node");
    oak::run_command_loop(dispatcher, grpc_channel.iter());
});