Macro oak::entrypoint_command_handler

macro_rules! entrypoint_command_handler {
    ($name : ident => $handler : ty) => { ... };
}

Similar to entrypoint, but offers a less verbose API in case in which the handler is a CommandHandler instance. In this case, the type of incoming messages is inferred from the generic type associated with the handler instance, and only the entrypoint name needs to be specified.

The generated implementation of the body of this node sets up logging, creates an instance of the handler, and then immediately calls run_command_loop on this instance.

It also automatically implements the WasmEntrypoint trait for the provided handler type which reflect the entrypoint name given via the macro invocation.

use oak_abi::proto::oak::application::ConfigMap;

#[derive(Default)]
struct Main;

impl oak::CommandHandler for Main {
    type Command = ConfigMap;
    fn handle_command(&mut self, command: ConfigMap) -> anyhow::Result<()> {
        // ...
    }
}

oak::entrypoint_command_handler!(oak_main => Main);