Module oak::proto::oak::application

Modules

node_configuration

Nested message and enum types in NodeConfiguration.

Structs

ApplicationConfiguration

An ApplicationConfiguration represents a unit of deployment in Oak.

ConfigMap

A serialized list of key-value pairs that are specified as command line flags to the Oak Loader binary, and are made available to the initial Node of the running Oak Application.

CryptoConfiguration

CryptoConfiguration describes the configuration of a cryptographic pseudo-Node (which is provided by the Oak Runtime).

GrpcClientConfiguration

GrpcClientConfiguration describes the configuration of a gRPC client pseudo-Node (which is provided by the Oak Runtime), connected to a specific external (non-Oak) gRPC service.

GrpcServerConfiguration

GrpcServerConfiguration describes the configuration of a gRPC server pseudo-Node (which is provided by the Oak Runtime), that processes gRPC requests from external (non-Oak) clients.

HttpClientConfiguration

HttpClientConfiguration describes the configuration of an HTTP client pseudo-Node (which is provided by the Oak Runtime), connected to a specific external (non-Oak) HTTP or HTTPS service.

HttpServerConfiguration

HttpServerConfiguration describes the configuration of an HTTP server pseudo-Node (which is provided by the Oak Runtime), that processes HTTP requests from external (non-Oak) clients.

LogConfiguration

LogConfiguration describes the configuration of a logging pseudo-Node (which is provided by the Oak Runtime).

ModuleSignature

Bundled signature of the SHA-256 hash of a Wasm module. TODO(#1830): Consider merging this with the SignedChallenge proto.

NodeConfiguration

NodeConfiguration indicates the configuration of a created Node.

RoughtimeClientConfiguration

RoughtimeClientConfiguration describes the configuration of a Roughtime client pseudo-Node (which is provided by the Oak Runtime), with the given external Roughtime servers and connection parameters.

RoughtimeServer

Information to identify a particular Roughtime server. Only UDP and Ed25519 public keys are currently supported.

StorageProxyConfiguration

StorageProxyConfiguration describes the configuration of a storage proxy pseudo-Node (which is provided by the Oak Runtime), connected to a specific storage provider.

WebAssemblyConfiguration

WebAssemblyConfiguration describes the configuration of a Web Assembly based Node.