Module oak::proto::oak::invocation

Structs

GrpcInvocation

A gRPC invocation, consisting of exactly two channels: one to read incoming requests from the client (wrapped in a Receiver), and one to write outgoing responses to the client (wrapped in a Sender).

GrpcInvocationReceiver

A wrapper for receiving a GrpcInvocation from a channel.

GrpcInvocationSender

A wrapper for sending a GrpcInvocation over a channel.

HttpInvocation

An HTTP invocation, consisting of exactly two channels: one to read incoming requests from the client (wrapped in a Receiver), and one to write outgoing responses to the client (wrapped in a Sender).

HttpInvocationSender

A wrapper for sending an HttpInvocation over a channel.

HttpInvocationSource

Source of an HTTP invocation, consisting of exactly two channels: one to send requests (wrapped in a Sender), and one to read the responses (wrapped in a Receiver).