Module oak::proto::oak::label

Modules

tag

Structs

GrpcTag
Policies related to gRPC communication, referring to the native gRPC node within the TCB.

Label
Label represents information flowing through a Node or channel.

Tag
Tag represents a category of confidentiality or integrity that is associated with data within Oak, and refers to a Node or family of Nodes which are able to declassify data with that tag.

TlsEndpointTag
Policies related to HTTPS communication.

WebAssemblyModuleTag
Policies related to modules, referring to the native WebAssembly node within the TCB.