Trait oak::CommandHandler

pub trait CommandHandler {
    type Command;
    fn handle_command(&mut self, command: Self::Command) -> Result<()>;
}

Trait implemented by Oak Nodes (or parts thereof) that operate on commands.

It has a single method for handling commands, which are usually received via the single incoming channel handle which is passed in at Node creation time, or derived from such stream.

Associated Types

type Command

Type of the command that is handled by this instance.

Required methods

fn handle_command(&mut self, command: Self::Command) -> Result<()>

Handles a single command instance.

The return value is only used for logging in case of error.