Trait oak::Node

pub trait Node<T: Decodable> {
    fn new() -> Self;

    fn handle_command(&mut self, command: T) -> Result<(), OakError>;
}

Trait implemented by all Oak Nodes.

It has a single method for handling commands, which are Decodable objects that are received via the single incoming channel handle which is passed in at node creation time. The return value is only used for logging in case of failure.

Required methods

fn new() -> Self

fn handle_command(&mut self, command: T) -> Result<(), OakError>

Loading content...