Trait oak::WithInit

pub trait WithInit {
    type Init;
    fn create(init: Self::Init) -> Self;
}

Trait implemented by structs that require an initialization step by receiving data over the inbound channel (usually wrapped within a [oak_io::InitWrapper] message).

Associated Types

type Init

Loading content...
Required methods

fn create(init: Self::Init) -> Self

Creates an instance of this struct based on the provided initialization data.

Panics if there is any error.

Loading content...