Crate oak

Modules

grpc
Functionality to help Oak Nodes interact with gRPC.

io
Wrappers for Oak SDK types to allow their use with std::io.

logger
A logger that sends output to an Oak logging channel, for use with the log facade.

node_config
Helper methods for creating common NodeConfiguration instances.

proto

rand
Functionality to allow use of the rand crate in Oak.

roughtime
Helper library for accessing the Oak Roughtime service.

storage
Helper library for accessing Oak storage services.

Macros

entrypoint
Register a new Node entrypoint.

Structs

Handle
Handle used to identify read or write channel halves.

Label
Label represents information flowing through a Node or channel.

ReadHandle
Wrapper for a handle to the read half of a channel.

WriteHandle
Wrapper for a handle to the send half of a channel.

Enums

ChannelReadStatus
Single byte values used to indicate the read status of a channel on the oak.wait_on_channels host function.

OakError
Generic Oak error.

OakStatus
Status values exchanged as i32 values across the Node Wasm interface.

Traits

Node
Trait implemented by all Oak Nodes.

Functions

channel_close
Close the specified channel Handle.

channel_create
Similar to channel_create_with_label, but with a fixed label corresponding to "public trusted".

channel_create_with_label
Create a new unidirectional channel.

channel_read
Read a message from a channel without blocking.

channel_write
Write a message to a channel.

node_create
Similar to node_create_with_label, but with a fixed label corresponding to "public trusted".

node_create_with_label
Creates a new Node running the configuration identified by config_name, running the entrypoint identified by entrypoint_name (for a Web Assembly Node; this parameter is ignored when creating a pseudo-Node), with the provided label, and passing it the given handle.

random_get
Fill a buffer with random data.

result_from_status
Convert a status returned from a host function call to a Result.

run_event_loop
Run an event loop on the provided node:

set_panic_hook
Install a panic hook that logs panic information.

wait_on_channels
Wait for one or more of the provided handles to become ready for reading from. On success, the returned vector of ChannelReadStatus values will be in 1-1 correspondence with the passed-in vector of Handles.