Struct oak::crypto::Crypto

pub struct Crypto { /* fields omitted */ }

Local representation of the connection to an external crypto service.

Implementations

impl Crypto

pub fn new(config: &CryptoConfiguration, label: &Label) -> Option<Crypto>

Creates a Crypto instance using the given configuration and label.

pub fn generate(&self, template: KeyTemplate) -> Result<KeysetHandle>

Generate a new keyset with a single fresh key that matches a specified template.

pub fn generate_named(&self, template_name: &str) -> Result<KeysetHandle>

Generate a new keyset with a single fresh key that matches a template specified by name.

pub fn public(&self, kh: KeysetHandle) -> Result<KeysetHandle>

Generate a new keyset that just holds public key information derived from an existing keyset that has private key information.

pub fn bind(
&self,
kh: KeysetHandle,
inner_kh: KeysetHandle,
format: KeysetFormat
) -> Result<Vec<u8>>

Convert a keyset to an encrypted form using another keyset.

pub fn unbind(
&self,
kh: KeysetHandle,
encrypted_keyset: &[u8],
format: KeysetFormat
) -> Result<KeysetHandle>

Retrieve a keyset from an encrypted form using another keyset.

pub fn kms_proxy(&self, kms_id: &str) -> Result<KeysetHandle>

Return a keyset that acts as a proxy to key material held in an external KMS.

pub fn encrypt(
&self,
kh: KeysetHandle,
pt: &[u8],
aad: &[u8]
) -> Result<Vec<u8>>

AEAD encryption.

pub fn decrypt(
&self,
kh: KeysetHandle,
ct: &[u8],
aad: &[u8]
) -> Result<Vec<u8>>

AEAD decryption.

pub fn encrypt_deterministically(
&self,
kh: KeysetHandle,
pt: &[u8],
aad: &[u8]
) -> Result<Vec<u8>>

Deterministic AEAD encryption.

pub fn decrypt_deterministically(
&self,
kh: KeysetHandle,
ct: &[u8],
aad: &[u8]
) -> Result<Vec<u8>>

Deterministic AEAD decryption.

pub fn compute_mac(&self, kh: KeysetHandle, data: &[u8]) -> Result<Vec<u8>>

Compute MAC.

pub fn verify_mac(
&self,
kh: KeysetHandle,
data: &[u8],
mac: &[u8]
) -> Result<()>

Verify MAC.

pub fn compute_prf(
&self,
kh: KeysetHandle,
data: &[u8],
output_length: u64
) -> Result<Vec<u8>>

Compute PRF.

pub fn sign(&self, kh: KeysetHandle, data: &[u8]) -> Result<Vec<u8>>

Generate signature.

pub fn verify(&self, kh: KeysetHandle, data: &[u8], sig: &[u8]) -> Result<()>

Verify signature.