Trait oak::grpc::Invocable

pub trait Invocable<T> {
    fn invoke(&self, target: &mut T, req: &[u8], writer: ChannelResponseWriter);
}

Trait implemented by all server method variants, applied to a server receiver instance having concrete type T.

Required methods

fn invoke(&self, target: &mut T, req: &[u8], writer: ChannelResponseWriter)

Generic function that handles request deserialization and response serialization (and sending down a channel) for a function that handles a request / response pair (either unary or streaming).

Panics if [de-]serialization or channel operations fail.

Loading content...