Trait oak::handle::HandleVisit

pub trait HandleVisit {
    fn visit<F>(&mut self, visitor: F) -> F
    where
        F: FnMut(&mut u64);
}

Visit all handles present in a type.

This trait can be automatically derived:

use oak_io::{handle::HandleVisit, Handle};

#[derive(HandleVisit)]
struct Thing {
    handle: Handle,
}

Alternatively, you can implement the trait manually. This is required if you want the trait to visit handles that are directly contained in a type.

use oak_io::{handle::HandleVisit, Handle};

struct Thing {
    handle: Handle,
}

impl HandleVisit for Thing {
    fn visit<F: FnMut(&mut Handle)>(&mut self, mut visitor: F) -> F {
        visitor(&mut self.handle);
        visitor
    }
}

Required methods

fn visit<F>(&mut self, visitor: F) -> F where
F: FnMut(&mut u64),

Invokes the provided closure on every handle contained in self.

The mutable reference allows modifying the handles.

Loading content...