Module oak::io

Wrappers for Oak SDK types to allow their use with std::io.

Structs

Message
A simple holder for bytes + handles, using internally owned buffers.

Receiver
Wrapper for a handle to the read half of a channel, allowing to receive data that can be decoded as bytes + handles via the Decodable trait.

Sender
Wrapper for a handle to the send half of a channel, allowing to send data that can be encoded as bytes + handles via the Encodable trait.

Traits

Decodable
A trait for objects that can be decoded from bytes + handles.

Encodable
A trait for objects that can be encoded as bytes + handles.

Functions

error_from_nonok_status
Map a non-OK OakStatus value to the nearest available std::io::Error.