Trait oak::io::ReceiverExt

pub trait ReceiverExt<T: Decodable> {
    fn close(&self) -> Result<(), OakStatus>;

    fn receive(&self) -> Result<T, OakError>;

    fn try_receive(&self) -> Result<T, OakError>;

    fn label(&self) -> Result<Label, OakError>;

    fn wait(&self) -> Result<ChannelReadStatus, OakStatus>;

    fn iter(self) -> ReceiverIterator<T>;
}

SDK-specific functionality provided by a receiver.

Required methods

fn close(&self) -> Result<(), OakStatus>

Closes the underlying channel used by the receiver.

fn receive(&self) -> Result<T, OakError>

Attempts to wait for a value on the receiver, blocking if necessary.

fn try_receive(&self) -> Result<T, OakError>

Attempts to read a value from the receiver, without blocking.

fn label(&self) -> Result<Label, OakError>

Retrieves the label associated with the underlying channel.

fn wait(&self) -> Result<ChannelReadStatus, OakStatus>

Waits for a value to be available from the receiver.

Returns ChannelReadStatus of the wrapped handle, or Err(OakStatus::ErrTerminated) if the Oak Runtime is terminating.

fn iter(self) -> ReceiverIterator<T>

Consumes the receiver and returns an iterator that will block waiting for messages. It will return None when the node terminates, or a non-transient error occurs when reading from the channel. Transient errors will be silently ignored by this iterator, including cases in which the incoming message cannot be decoded correctly.

Also see std::sync::mpsc::Receiver::iter.

Loading content...