Macro oak::entrypoint

macro_rules! entrypoint {
    ($name:ident => $node:expr) => { ... };
}

Register a new Node entrypoint.

This registers the entrypoint name and the expression used to construct the Node instance. The returned object should implement the Node trait.

#[derive(Default)]
struct DummyNode;

impl oak::Node<DummyCommand> for DummyNode {
    // ...
}

oak::entrypoint!(dummy => DummyNode::default());

Entrypoints need to be declared in the global scope (as opposed to for example in a function body). Is recommended but not required to define all entrypoints of your module in lib.rs.

If instantiating your Node requires some setup, it is possible to do that in the Node expression too:

oak::entrypoint!(its_complicated => {
    init_all_the_things();
    DummyNode::default()
});