Trait oak::WasmEntrypoint

pub trait WasmEntrypoint {
    type Message: Encodable + Decodable;

    const ENTRYPOINT_NAME: &'static str;
}

Trait implemented by structs that correspond to entrypoint of Wasm nodes.

This statically ties together the Wasm entrypoint name (i.e. the name of the Wasm exported function) and the type of messages that the node accepts with this trait.

This trait is automatically implemented for types used as part of entrypoint_command_handler entrypoint definitions.

Associated Types

type Message: Encodable + Decodable

Type of message that the Node accepts on its inbound channel.

Loading content...
Associated Constants

const ENTRYPOINT_NAME: &'static str

Name of the WebAssembly exported function that corresponds to the entrypoint of this Node.

Loading content...